analytics-engineering/pyproject.toml
[project]
name = "analytics-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-core>=2.29.0",
    "google-cloud-storage>=3.9.0",
    "python-dotenv>=1.2.1",
]
